v938 * Preliminary support for PX4 firmware

+++ MinimOsd_Extra/Vars.h
@@ -99,7 +99,7 @@ static uint16_t     temperature = 0;  // temp*100
 
 
 static uint16_t      remaining_estimated_flight_time_seconds = 0;
-static uint8_t       osd_mode = 0;                   // Navigation mode from RC AC2 = CH5, APM = CH8
+static uint32_t      osd_mode = 0;  // 32 for PX4 compatibility        // Navigation mode from RC AC2 = CH5, APM = CH8
 
 static unsigned long timer_20ms = 0;
 static unsigned long timer_100ms = 0;
